refactor: extract editor init as custom hook
To improve readability and add more docs.
