+++ includes/publisher/js/src/components/RichText/RichTextEditor.jsx
@@ -1,13 +1,12 @@
-/* global atlasContentModelerFormEditingExperience */
-import React, { useEffect, useRef, useState } from "react";
+import React, { useState } from "react";
 import { v4 as uuidv4 } from "uuid";
 import SoloRichTextEditorField from "./SoloRichTextEditorField";
 import RepeatingRichTextEditorField from "./RepeatingRichTextEditorField";
+import useWpEditorInitialize from "./useWPEditorInitialize";
 import { __ } from "@wordpress/i18n";
-const { wp } = window;
 
 export default function RichTextEditor({ field, modelSlug }) {
-	const editorReadyTimer = useRef(null);
+	// Generates a unique ID for each rich text field for initialization and keying.
 	const initialValues = field?.isRepeatableRichText
 		? (field?.value || [""]).map((val) => {
 				return { id: uuidv4(), value: val };
@@ -16,48 +15,7 @@ export default function RichTextEditor({ field, modelSlug }) {
 
 	const [values, setValues] = useState(initialValues);
 
-	useEffect(() => {
-		const editorReadyTime = 500;
-		const initializeEditorWhenReady = () => {
-			/**
-			 * WP defines getDefaultSettings() in an admin footer script tag after
-			 * admin scripts are enqueued, so we must wait for it to be available.
-			 */
-			if (typeof wp?.oldEditor?.getDefaultSettings === "function") {
-				let editorsToInitialize = values.map(({ id }) => id);
-
-				editorsToInitialize.forEach((editorId) => {
-					wp.oldEditor.initialize(editorId, {
-						...wp.oldEditor.getDefaultSettings(),
-						tinymce: {
-							height: "300",
-							toolbar1:
-								"undo,redo,formatselect,bold,italic,bullist,numlist,blockquote,alignleft,aligncenter,alignright,link,unlink,wp_add_media",
-						},
-						mediaButtons: true,
-						quicktags: false,
-					});
-				});
-			} else {
-				editorReadyTimer.current = setTimeout(
-					initializeEditorWhenReady,
-					editorReadyTime
-				);
-			}
-		};
-
-		if (
-			atlasContentModelerFormEditingExperience?.models ||
-			atlasContentModelerFormEditingExperience?.models[
-				atlasContentModelerFormEditingExperience.postType
-			]
-		) {
-			initializeEditorWhenReady();
-		}
-		return () => {
-			clearTimeout(editorReadyTimer.current);
-		};
-	}, [editorReadyTimer, values]);
+	useWpEditorInitialize(values);
 
 	return field?.isRepeatableRichText ? (
 		<RepeatingRichTextEditorField
